+++ tp6/animations/Project3/animation.h
@@ -7,15 +7,15 @@ private:
 	char anim_id;
 	string anim_prefix;
 	unsigned int anim_cant_img;
-	float anim_period;
-	float speed;
+	double anim_period;
+	double speed;
 public:
 	animation(char name,string pref, unsigned int im_cant,double per, double vel);
 	
 	ALLEGRO_BITMAP ** p2anim_lib;
 	char& get_anim_id(void);
-	float& get_anim_per(void);
-	float& get_speed(void);
+	double& get_anim_per(void);
+	double& get_speed(void);
 	string& get_anim_prefix(void);
 	unsigned int& get_anim_cant_img(void);
 	void play_anim(char anim_type,int& i, float& k, float MAX_SCREEN_WIDTH, float MAX_SCREEN_HEIGHT, bool& sense, unsigned int img_qnt);
